GS/KW Created a Controller class
